requirements.txt
# Requirements for Indonesian Speech-to-Text with Whisper
# Python 3.9.9 (as specified in the Whisper README.md)

openai-whisper==20250625
torch==2.8.0
numba==0.60.0
numpy==2.0.2
tqdm==4.67.1
more-itertools==10.8.0
tiktoken==0.11.0
triton==3.4.0
typing-extensions==4.15.0
regex==2025.9.1
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
filelock==3.19.1
fsspec==2025.9.0
sympy==1.14.0
mpmath==1.3.0
networkx==3.2.1
jinja2==3.1.6
markupsafe==3.0.2
llvmlite==0.43.0
zipp==3.23.0
importlib-metadata==8.7.0

# Note: For GPU support with CUDA, you need:
# - CUDA-capable GPU
# - Appropriate NVIDIA drivers installed
# - ffmpeg installed on the system
flask==3.1.2
flask-socketio==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
bidict==0.23.1
simple-websocket==1.1.0
wsproto==1.2.0
h11==0.16.0
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
werkzeug==3.1.3
